// src/master-setting/interfaces/i-company-info.ts
export interface ICompanyInfo {
    // #region Properties (9)

    companyId: string;
    containerId: string;
    partnerId: string;
    tokenId: string;
    uid: string;
    userEmail: string;
    userName: string;
    userPhoneNumber: string;
    userPhotoUrl: string;

    // #endregion Properties (9)
}